split envsh and --watch into separate commands
